\name{followup}
\alias{followup}
\title{Followup data from experiment on soldered joints.}
\description{This is the followup data from the experiment analyzed in Culp, Ryan, Chen, and Hamada (2018). These data were first presented and described in de Mast and van Wieringen (2010). Thirty parts were rated on a 4-point ordinal scale of 1-4 by each of 3 raters in a balanced design of 2 repetitions per combination of part and rater.}
\usage{data("followup")}
\format{
  This data frame has 30 rows (i.e., one for each part) and the following 6 columns.
  \describe{
    \item{\code{Yi11}}{ordinal value assigned by rater 1 on repetition 1.}
    \item{\code{Yi12}}{ordinal value assigned by rater 1 on repetition 2.}
    \item{\code{Yi21}}{ordinal value assigned by rater 2 on repetition 1.}
    \item{\code{Yi22}}{ordinal value assigned by rater 2 on repetition 2.}
    \item{\code{Yi31}}{ordinal value assigned by rater 3 on repetition 1.}
    \item{\code{Yi32}}{ordinal value assigned by rater 3 on repetition 2.}
   }
}

\author{Ken Ryan}

\references{
	Culp, S.L., Ryan, K.J., Chen, J., and Hamada, M.S. (2018). \dQuote{Analysis of Repeatability and Reproducibility Studies with Ordinal Measurements.} \emph{Technometrics}, doi:10.1080/00401706.2018.1429317.
	
	de Mast, J. and van Wieringen, W.N. (2010). \dQuote{Modeling and Evaluating Repeatability and Reproducibility of Ordinal Classifications.} \emph{Technometrics}, 52(1), 94-106.
	
	
}

\keyword{datasets}
